+++ neo/Storage/Common/DataCache.py
@@ -88,12 +88,12 @@ class DataCache:
             key_prefix = b''
 
         for k, v in self.FindInternal(key_prefix):
-            if bytes(key_prefix + k) not in self.dictionary:
+            if k not in self.dictionary:
                 yield k, v
 
         for k, v in self.dictionary.items():
             if v.State != TrackState.DELETED and (key_prefix is None or key_prefix in k):
-                yield k[20:], v.Item
+                yield k, v.Item
 
     def FindInternal(self, key_prefix) -> dict:
         # should be equal to DBInterface.Find